refactored Byon IdCreation for use in matchmaking
